requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai>=0.5.0
pydantic==2.5.3
python-dotenv==1.0.0
playwright==1.40.0
playwright-stealth==2.0.0
beanie>=1.23.0
motor>=3.3.0
